% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_true_value.R
\name{calc_true_value}
\alias{calc_true_value}
\title{Calculate true value of PSEs.}
\usage{
calc_true_value(SC, intv = 3)
}
\arguments{
\item{SC}{A vector of parameters, in order of: b0, bw, bq, bs, a0, aw, aq, d0, dw, sigma_q, sigma_s.}

\item{intv}{Number of intervention, 3 or 4. Default is 3.}
}
\description{
Calculate true value of PSEs.
}
\examples{
calc_true_value(rep(0.5,11))
calc_true_value(rep(0.5,11), intv=4)
}
