% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_analysis.R
\name{mediation_analysis}
\alias{mediation_analysis}
\title{Apply mediation analysis for non-rare binary outcome with two continuous mediators}
\usage{
mediation_analysis(dt, confounders = c(), nb = 0, intv = 3)
}
\arguments{
\item{dt}{Input data.}

\item{confounders}{Confounder values.}

\item{nb}{Number of bootstrapping. Default is 0 (no bootstrapping applied).}

\item{intv}{Number of intervention, 3 or 4. Default is 3.}
}
\description{
Apply mediation analysis for non-rare binary outcome with two continuous mediators
}
\examples{
para=c(rep(-0.5,9),1,1)
dat1=sim_mediation_data(0.5,1000,para) #binary exposure
apply(dat1,2,mean) #the proportion of Y should not be too skew (nearly 0 or 1)
res1_3=mediation_analysis(dat1)
tru1_3=calc_true_value(para)
dat2=sim_mediation_data(c(0,1),1000,para) #continous exposure
res2_4=mediation_analysis(dat2, intv=4, nb=500)
tru2_4=calc_true_value(para, intv=4)
}
\keyword{Causal}
\keyword{Mediation}
\keyword{analysis,}
\keyword{inference.}
